+++ header/GraphicsItem/protocol.h
@@ -10,7 +10,7 @@
 struct Vertex
 {
     QPointF point;
-    float bulge;
+    double bulge;
 };
 
 enum LineType
@@ -27,8 +27,8 @@ enum ItemTypeId /* 只用于对照 不参与实际使用 */
 
 extern  QPen defaultLinePen;
 
-void getCircleFromBulge(const QPointF& p1, const QPointF& p2, const float & bulge, QPointF& center,float & radius);
-void getCircleFromThreePoints(const QPointF& p1, const QPointF& p2, const QPointF& p3, QPointF& center, float & radius);
-void getBulgeFromThreePoints(const QPointF& p1, const QPointF& p2, const QPointF& p3,const QPointF& center,const float & radius, float  &bulge);
-QPainterPath createArcPath(const QPointF& p1, const QPointF& p3, const float & bulge);
+void getCircleFromBulge(const QPointF& p1, const QPointF& p2, const double & bulge, QPointF& center,double & radius);
+void getCircleFromThreePoints(const QPointF& p1, const QPointF& p2, const QPointF& p3, QPointF& center, double & radius);
+void getBulgeFromThreePoints(const QPointF& p1, const QPointF& p2, const QPointF& p3,const QPointF& center,const double & radius, double  &bulge);
+QPainterPath createArcPath(const QPointF& p1, const QPointF& p3, const double & bulge);
 #endif // PROTOCOL_H
